@@ -13,23 +13,24 @@
         "sort-packages": true
     },
     "require": {
-        "php": "^7.2",
+        "php": "^7.4 || ^8.0",
         "ext-pcntl": "*",
-        "ext-rdkafka": "^4",
+        "ext-rdkafka": "^4 || ^5",
         "psr/log": "^1.1"
     },
     "require-dev": {
-        "doctrine/coding-standard": "^7.0",
-        "infection/infection": "^0.16.3",
+        "doctrine/coding-standard": "^8.2",
+        "infection/infection": "^0.20.2",
         "kwn/php-rdkafka-stubs": "^2.0",
-        "phpstan/extension-installer": "^1.0",
-        "phpstan/phpstan": "0.12.25",
-        "phpstan/phpstan-phpunit": "0.12.8",
-        "phpstan/phpstan-strict-rules": "0.12.2",
-        "phpunit/phpunit": "^9.1",
-        "psalm/plugin-phpunit": "^0.10.0",
-        "simpod/php-coveralls-mirror": "^3.0",
-        "vimeo/psalm": "^3.12"
+        "phpstan/extension-installer": "^1.1",
+        "phpstan/phpstan": "0.12.68",
+        "phpstan/phpstan-phpunit": "0.12.17",
+        "phpstan/phpstan-strict-rules": "0.12.9",
+        "phpunit/phpunit": "^9.2",
+        "psalm/plugin-phpunit": "^0.15.0",
+        "roave/infection-static-analysis-plugin": "^1.7",
+        "thecodingmachine/phpstan-safe-rule": "^1.0",
+        "vimeo/psalm": "^4.4"
     },
     "suggest": {
         "kwn/php-rdkafka-stubs": "Support and autocompletion for RDKafka in IDE | require as dev dependency"
