@@ -8,15 +8,14 @@ use DateTimeImmutable;
 
 final class BatchTime
 {
-    /** @var int */
-    public $endMsTimestamp;
+    public int $endMsTimestamp;
 
     public function __construct(int $timeoutMs, DateTimeImmutable $now)
     {
         $this->reset($timeoutMs, $now);
     }
 
-    public function reset(int $timeoutMs, DateTimeImmutable $now) : void
+    public function reset(int $timeoutMs, DateTimeImmutable $now): void
     {
         $this->endMsTimestamp = $now->getTimestamp() * 1000 + $timeoutMs;
     }
