@@ -8,7 +8,7 @@ use SimPod\Kafka\Common\Exception\KafkaException;
 
 final class RecordNotFound extends KafkaException
 {
-    public static function setEmpty() : self
+    public static function setEmpty(): self
     {
         return new self('Record was not found because record set is empty');
     }
