@@ -10,7 +10,7 @@ use SimPod\Kafka\Clients\Consumer\Exception\IncompatibleStatus;
 
 final class IncompatibleStatusTest extends TestCase
 {
-    public function testFromMessage() : void
+    public function testFromMessage(): void
     {
         $message      = new Message();
         $message->err = 1;
@@ -18,6 +18,7 @@ final class IncompatibleStatusTest extends TestCase
 
         self::assertSame(
             'Consumer status "1" is not handled: Broker: Offset out of range',
-            $exception->getMessage());
+            $exception->getMessage()
+        );
     }
 }
